Fix build and moved TextFileReader #1105
- fixed dependencies and build setup
- moved the TextFileReader from src/test/java to src/main/java, so it
can be used with the sechub-commons-model dependency
